requirements.txt
tkinter
Pillow
numpy
scikit-image
matplotlib
pandas
seaborn
opencv-python
pydicom
tensorflow
keras